Update backend directory structure and improve data retrieval logic; apply various minor fixes and optimizations.
