+++ parkulator-backend/src/main/java/hr/parkulator/parkulator_backend/controller/ParkingController.java
@@ -1,7 +1,7 @@
 package hr.parkulator.parkulator_backend.controller;
 
+import hr.parkulator.parkulator_backend.dto.parking.ParkingDTO;
 import hr.parkulator.parkulator_backend.entities.Parking;
-import hr.parkulator.parkulator_backend.dto.ParkingDTO;
 import hr.parkulator.parkulator_backend.services.ParkingService;
 import org.springframework.web.bind.annotation.*;
