--- parkulator-backend/src/main/java/hr/parkulator/parkulator_backend/controller/TestController.java
@@ -1,28 +1,0 @@
-package hr.parkulator.parkulator_backend.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import hr.parkulator.parkulator_backend.services.ParkingDataService;
-
-
-@RestController
-public class TestController {
-    ParkingDataService parkingDataService;
-
-    public TestController(ParkingDataService parkingDataService) {
-        this.parkingDataService = parkingDataService;
-    }
-   
-    @GetMapping("/test-live-data")
-    public String testLiveData() {
-        parkingDataService.saveInitialData();
-        return "POSLANO";
-    }
-    @GetMapping("/refresh")
-    public String refresh() {
-        parkingDataService.saveRefreshData();
-        return "Refreshano";
-    }
-    
-}
