--- parkulator-backend/src/main/java/hr/parkulator/parkulator_backend/dto/ParkingRefreshDTO.java
@@ -1,16 +1,0 @@
-package hr.parkulator.parkulator_backend.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-@AllArgsConstructor
-public class ParkingRefreshDTO {
-    private String sourceKey;
-    private String name;
-    private boolean isLive;
-    private Long availableSpots;
-    private List<ParkingPriceDTO> parkingPrice;
-}
