+++ parkulator-backend/src/main/java/hr/parkulator/parkulator_backend/dto/parking/ParkingDTO.java
@@ -1,4 +1,4 @@
-package hr.parkulator.parkulator_backend.dto;
+package hr.parkulator.parkulator_backend.dto.parking;
 
 import lombok.*;
