+++ parkulator-backend/src/main/java/hr/parkulator/parkulator_backend/dto/parking/ParkingRefreshDTO.java
@@ -1,4 +1,5 @@
-package hr.parkulator.parkulator_backend.dto;
+package hr.parkulator.parkulator_backend.dto.parking;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
@@ -6,14 +7,10 @@ import java.util.List;
 
 @Data
 @AllArgsConstructor
-public class ParkingDataDTO {
+public class ParkingRefreshDTO {
     private String sourceKey;
     private String name;
-    private String address;
-    private String link;
-    private String type;
     private boolean isLive;
-    private Long spots;
     private Long availableSpots;
-    List<ParkingPriceDTO> parkingPrice;
+    private List<ParkingPriceDTO> parkingPrice;
 }
